Moved IDataPluginServices, stubs and getIndexPatternFromFilter to NP
