+++ src/plugins/data/public/query/filter_manager/index.ts
@@ -23,3 +23,5 @@ export { uniqFilters } from './lib/uniq_filters';
 export { mapAndFlattenFilters } from './lib/map_and_flatten_filters';
 export { onlyDisabledFiltersChanged } from './lib/only_disabled';
 export { generateFilters } from './lib/generate_filters';
+export { getDisplayValueFromFilter } from './lib/display/get_display_value';
+export { getIndexPatternFromFilter } from './lib/display/get_index_pattern_from_filter';
