+++ src/plugins/data/public/query/filter_manager/lib/display/get_display_value.ts
@@ -18,24 +18,22 @@
  */
 
 import { get } from 'lodash';
-import { esFilters } from '../../../../../../../../plugins/data/public';
-import { IndexPattern } from '../../../../index_patterns/index_patterns';
-import { Field } from '../../../../index_patterns/fields';
-import { getIndexPatternFromFilter } from './filter_editor_utils';
+import { esFilters, IIndexPattern, IFieldType } from '../../../..';
+import { getIndexPatternFromFilter } from './get_index_pattern_from_filter';
 
-function getValueFormatter(indexPattern?: IndexPattern, key?: string) {
+function getValueFormatter(indexPattern?: IIndexPattern, key?: string) {
   if (!indexPattern || !key) return;
   let format = get(indexPattern, ['fields', 'byName', key, 'format']);
-  if (!format && indexPattern.fields.getByName) {
+  if (!format && (indexPattern.fields as any).getByName) {
     // TODO: Why is indexPatterns sometimes a map and sometimes an array?
-    format = (indexPattern.fields.getByName(key) as Field).format;
+    format = ((indexPattern.fields as any).getByName(key) as IFieldType).format;
   }
   return format;
 }
 
 export function getDisplayValueFromFilter(
   filter: esFilters.Filter,
-  indexPatterns: IndexPattern[]
+  indexPatterns: IIndexPattern[]
 ): string {
   const indexPattern = getIndexPatternFromFilter(filter, indexPatterns);
